Add autogeneration of grid's width and height
Now grid width and height are generated based on the size of the window.
Grid dimensions stay the same if the window size is changed. But if the
reset button is clicked, then grid dimensions change based on the current
window size. Reset button doesn't work when the game is active.